requirements.txt
pip
setuptools
wheel

python-dotenv
requests

fastapi

pydantic

uvicorn

python-jose

bcrypt

tinydb

httpx

pytest